Add flags that control building the executables
The `readme` and `play-tomland` have very limited usefulness for usual
package users. Therefore it is better to disable them by default in
order to conserve resources.

Additionally, because `build-tool-depends` as used in the `readme`
executable are currently unsupported by Hackage's docs builder,
automated docs builds didn't work for `tomland` and any packages that
depend on it.

Context:
https://github.com/haskell/hackage-server/issues/988#issuecomment-956508617